+++ projects/client/audio/audio_player/source/audio/Speaker.h
@@ -15,7 +15,7 @@ namespace audio {
       virtual Device_Settings start() = 0;
       virtual void stop() = 0;
 
-      void update(uint8_t *data, int length);
+      void update(float *data, int length);
 
       void set_player(Player *player) {
         this->player = player;
